model/pyproject.toml
[project]
name = "mina-model"
version = "0.1.0"
description = "YOLOv8 model for fish disease detection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "roboflow>=1.2.11",
    "ultralytics>=8.3.0",
    "torch>=2.0.0",
    "hypothesis>=6.0.0",
    "pytest>=8.0.0",
    "numpy==2.0.2",
    "scipy==1.13.1",
    "tensorflow==2.18.0",
    "pillow>=10.0.0",
    "kaggle>=1.6.0",
    "onnxruntime>=1.17.0",
    "onnxscript>=0.1.0",
    "onnx>=1.14.0,<1.17.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
mina-train = "cli.train:main"
mina-export = "cli.export:main"
mina-evaluate = "cli.evaluate:main"
mina-infer = "cli.infer:main"
mina-download = "cli.download:main"
mina-tune = "cli.tune:main"
# Gate classifier
gate-download  = "cli.gate_download:main"
gate-train     = "cli.gate_train:main"
gate-export    = "cli.gate_export:main"
gate-evaluate  = "cli.gate_evaluate:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mina", "cli"]
